requirements.txt
# Web UI
streamlit==1.25.0

# PDF / DOCX / TXT extraction
pdfplumber==0.9.0
python-docx==0.8.11

# Text to speech
gTTS==2.3.1
pydub==0.25.1

# Google Gemini LLM
google-generativeai==0.4.0

# Environment variable handling
python-dotenv==1.1.1

# Optional utilities
numpy==1.25.2
pandas==2.1.0

# FFmpeg wrapper (required for pydub to handle MP3/WAV)
ffmpeg-python